Update garyutils
added combineTableElements... fun

usage: give a table and separation character, then, the function will loop over every item in the given table, adding it to a return table if it has the character at the start and end, or combining it with each consecutive item until finding an item with the separation character. returns the table with the combined characters, removing the seperation characters from the items.